[feat] Add plugin, change colorscheme and Shortcuts
It's hard to find the perfect colorscheme. I've changed the monitor
position so I needed to change the colorscheme lol. Anyways,
vim-startify is cool. The matches show in the center of screen. As I'm
use the quickfix list a lot to "find and replace", I needed to use
shortcuts to loop over the list